pyproject.toml
[project]
name = "adk-github"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk==1.3.0",
]
